requirements_backup_20260311.txt
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
curl_cffi==0.13.0
defusedxml==0.7.1
filelock==3.20.0
Flask==3.1.2
frozendict==2.4.7
fsspec==2025.12.0
futu_api==10.0.6008
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
multitasking==0.0.12
networkx==3.6.1
numpy==2.4.2
pandas==3.0.1
peewee==4.0.1
platformdirs==4.9.4
protobuf==3.20.3
pycparser==3.0
pycryptodome==3.23.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
setuptools==70.2.0
simplejson==3.20.2
six==1.17.0
soupsieve==2.8.3
sympy==1.14.0
torch==2.10.0+cpu
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
Werkzeug==3.1.5
wheel==0.46.3
yfinance==1.2.0
youtube-transcript-api==1.2.4
